requirements.txt
fastapi==0.77.1
pydantic==1.9.0
uvicorn==0.17.6
requests==2.27.1
passlib==1.7.4
sqlalchemy==1.4.36
alembic==1.7.7
python-jose==3.3.0
pymysql==1.0.2
bcrypt==3.2.2
cryptography==37.0.2